Optimize more

+++ include/battle_setup.h
@@ -62,6 +62,6 @@ u16 GetLastBeatenRematchTrainerId(u16 trainerId);
 bool8 ShouldTryRematchBattle(void);
 bool8 IsTrainerReadyForRematch(void);
 void ShouldTryGetTrainerScript(void);
-u16 CountBattledRematchTeams(u16 trainerId);
+u32 CountBattledRematchTeams(u16 trainerId);
 
 #endif // GUARD_BATTLE_SETUP_H
